+++ src/Exceptions/ArgumentPropagateException.php
@@ -1,0 +1,68 @@
+<?php
+
+namespace ArrayFunctions\Exceptions;
+
+use Message;
+use MWException;
+use Throwable;
+
+/**
+ * Thrown when a parser function was an error as an argument.
+ */
+class ArgumentPropagateException extends MWException {
+	/**
+	 * @var int|string The name of the argument that was passed the error
+	 */
+	private $argName;
+
+	/**
+	 * @var string The ID of the error
+	 */
+	private string $errorId;
+
+	/**
+	 * @var Message The message of the error
+	 */
+	private Message $errorMessage;
+
+	/**
+	 * @param string|int $argName
+	 * @param string $errorId
+	 * @param Message $errorMessage
+	 * @param Throwable|null $previous
+	 */
+	public function __construct( $argName, string $errorId, Message $errorMessage, ?Throwable $previous = null ) {
+		parent::__construct( '', 0, $previous );
+
+		$this->argName = $argName;
+		$this->errorId = $errorId;
+		$this->errorMessage = $errorMessage;
+	}
+
+	/**
+	 * Return the ame of the argument that was passed the error.
+	 *
+	 * @return int|string
+	 */
+	public function getArgName() {
+		return $this->argName;
+	}
+
+	/**
+	 * Return the ID of the error.
+	 *
+	 * @return string
+	 */
+	public function getErrorId(): string {
+		return $this->errorId;
+	}
+
+	/**
+	 * Return the message of the error.
+	 *
+	 * @return Message
+	 */
+	public function getErrorMessage(): Message {
+		return $this->errorMessage;
+	}
+}