+++ src/ParserInitHookHandler.php
@@ -2,6 +2,7 @@
 
 namespace ArrayFunctions;
 
+use Config;
 use MediaWiki\Hook\ParserFirstCallInitHook;
 use MWException;
 use Parser;
@@ -21,12 +22,18 @@ class ParserInitHookHandler implements ParserFirstCallInitHook {
 	private ArrayFunctionRegistry $registry;
 
 	/**
+	 * @var Config
+	 */
+	private Config $config;
+
+	/**
 	 * @param ArrayFunctionFactory $factory
 	 * @param ArrayFunctionRegistry $registry
 	 */
-	public function __construct( ArrayFunctionFactory $factory, ArrayFunctionRegistry $registry ) {
+	public function __construct( ArrayFunctionFactory $factory, ArrayFunctionRegistry $registry, Config $config ) {
 		$this->factory = $factory;
 		$this->registry = $registry;
+		$this->config = $config;
 	}
 
 	/**
@@ -34,17 +41,19 @@ class ParserInitHookHandler implements ParserFirstCallInitHook {
 	 * @throws MWException
 	 */
 	public function onParserFirstCallInit( $parser ): void {
+		$enableErrorTracking = $this->config->get( 'ArrayFunctionsEnableErrorTracking' );
+
 		foreach ( $this->registry->getFunctions() as $class ) {
 			$parser->setFunctionHook(
 				$class::getName(),
-				[ new ArrayFunctionInvoker( $class, $this->factory ), "invoke" ],
+				[ new ArrayFunctionInvoker( $class, $enableErrorTracking, $this->factory ), "invoke" ],
 				Parser::SFH_OBJECT_ARGS
 			);
 
 			foreach ( $class::getAliases() as $alias ) {
 				$parser->setFunctionHook(
 					$alias,
-					[ new ArrayFunctionInvoker( $class, $this->factory ), "invoke" ],
+					[ new ArrayFunctionInvoker( $class, $enableErrorTracking, $this->factory ), "invoke" ],
 					Parser::SFH_OBJECT_ARGS
 				);
 			}